Separating the propagators
